refactored styles into sepparate files and added the return of the project to the admin report
